content/canvas/test/test_2d.imageData.get.sourcesize.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.get.sourcesize</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var imgdata1 = ctx.getImageData(0, 0, 10, 10);
var imgdata2 = ctx.getImageData(0, 0, 20, 20);
ok(imgdata2.width > imgdata1.width, "imgdata2.width > imgdata1.width");
ok(imgdata2.height > imgdata1.height, "imgdata2.height > imgdata1.height");

if (!_deferred) SimpleTest.finish();
});
</script>
